Renaming of all question files
