+++ edsl/questions/__init__.py
@@ -3,26 +3,25 @@ from edsl.questions.settings import Settings
 from edsl.questions.register_questions_meta import RegisterQuestionsMeta
 
 # Base Class
-from .QuestionBase import QuestionBase
+from .question_base import QuestionBase
 
 # Core Questions
-from .QuestionCheckBox import QuestionCheckBox
-from .QuestionExtract import QuestionExtract
-from .QuestionFreeText import QuestionFreeText
-from .QuestionFunctional import QuestionFunctional
-from .QuestionList import QuestionList
-from .QuestionMatrix import QuestionMatrix
-from .QuestionDict import QuestionDict
-from .QuestionMultipleChoice import QuestionMultipleChoice
-from .QuestionNumerical import QuestionNumerical
-from .QuestionBudget import QuestionBudget
-from .QuestionRank import QuestionRank
+from .question_check_box import QuestionCheckBox
+from .question_extract import QuestionExtract
+from .question_free_text import QuestionFreeText
+from .question_functional import QuestionFunctional
+from .question_list import QuestionList
+from .question_matrix import QuestionMatrix
+from .question_dict import QuestionDict
+from .question_multiple_choice import QuestionMultipleChoice
+from .question_numerical import QuestionNumerical
+from .question_budget import QuestionBudget
+from .question_rank import QuestionRank
 
 # # # Questions derived from core questions
-from .derived.QuestionLikertFive import QuestionLikertFive
-from .derived.QuestionLinearScale import QuestionLinearScale
-from .derived.QuestionYesNo import QuestionYesNo
-from .derived.QuestionTopK import QuestionTopK
-
+from .derived.question_likert_five import QuestionLikertFive
+from .derived.question_linear_scale import QuestionLinearScale
+from .derived.question_yes_no import QuestionYesNo
+from .derived.question_top_k import QuestionTopK
 # # Compose Questions
 # from edsl.questions.compose_questions import compose_questions
