+++ edsl/questions/derived/question_likert_five.py
@@ -1,8 +1,7 @@
 from __future__ import annotations
 from typing import Optional
-from edsl.questions.QuestionMultipleChoice import QuestionMultipleChoice
-
-from edsl.questions.decorators import inject_exception
+from ..question_multiple_choice import QuestionMultipleChoice
+from ..decorators import inject_exception
 
 
 class QuestionLikertFive(QuestionMultipleChoice):
