+++ edsl/questions/derived/question_yes_no.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 from typing import Optional
-from edsl.questions.descriptors import QuestionOptionsDescriptor
-from edsl.questions.QuestionMultipleChoice import QuestionMultipleChoice
+from ..descriptors import QuestionOptionsDescriptor
+from ..question_multiple_choice import QuestionMultipleChoice
 
-from edsl.questions.decorators import inject_exception
+from ..decorators import inject_exception
 
 
 class QuestionYesNo(QuestionMultipleChoice):
