+++ edsl/questions/question_extract.py
@@ -1,17 +1,17 @@
 from __future__ import annotations
 import json
 import re
-
-from typing import Any, Optional, Dict
-from edsl.questions.QuestionBase import QuestionBase
-from edsl.questions.descriptors import AnswerTemplateDescriptor
-
-from edsl.questions.response_validator_abc import ResponseValidatorABC
-from edsl.questions.data_structures import BaseResponse
-from edsl.questions.decorators import inject_exception
-
 from typing import Dict, Any
+from typing import Any, Optional, Dict
+
 from pydantic import create_model, Field
+
+from .question_base import QuestionBase
+from .descriptors import AnswerTemplateDescriptor
+
+from .response_validator_abc import ResponseValidatorABC
+from .data_structures import BaseResponse
+from .decorators import inject_exception
 
 
 def extract_json(text, expected_keys, verbose=False):
