+++ tests/prompts/test_adding_custom_prompts_to_questions.py
@@ -1,5 +1,5 @@
-from edsl.questions.QuestionMultipleChoice import QuestionMultipleChoice
-from edsl.language_models.model import Model
+from edsl.questions import QuestionMultipleChoice
+from edsl.language_models import Model
 
 new_instructions = """You are being asked the following question: {{question_text}}
     The options are
