+++ tests/questions/test_QuestionBudget.py
@@ -2,7 +2,7 @@ import pytest
 import uuid
 from edsl.exceptions.questions import QuestionAnswerValidationError
 from edsl.questions import QuestionBudget, Settings
-from edsl.questions.QuestionBudget import QuestionBudget, main
+from edsl.questions.question_budget import QuestionBudget, main
 
 
 # def test_QuestionBudget_main():
