+++ tests/questions/test_QuestionExtract.py
@@ -1,4 +1,4 @@
-from edsl.questions.QuestionExtract import main
+from edsl.questions.question_extract import main
 
 
 def test_QuestionExtract_main():
