+++ tests/questions/test_QuestionFreeText.py
@@ -4,8 +4,8 @@ from edsl.exceptions.questions import (
     QuestionResponseValidationError,
 )
 from edsl.questions import Settings
-from edsl.questions.QuestionBase import QuestionBase
-from edsl.questions.QuestionFreeText import QuestionFreeText, main
+from edsl.questions import QuestionBase
+from edsl.questions.question_free_text import QuestionFreeText, main
 
 
 def test_QuestionFreeText_main():
