+++ tests/questions/test_QuestionLinearScale.py
@@ -3,8 +3,8 @@ from edsl.exceptions.questions import (
     QuestionAnswerValidationError,
     QuestionCreationValidationError,
 )
-from edsl.questions.QuestionBase import QuestionBase
-from edsl.questions.derived.QuestionLinearScale import QuestionLinearScale, main
+from edsl.questions.question_base import QuestionBase
+from edsl.questions.derived.question_linear_scale import QuestionLinearScale, main
 
 
 def test_QuestionLinearScale_main():
