+++ tests/questions/test_QuestionMatrix.py
@@ -1,7 +1,7 @@
 import pytest
 from edsl.exceptions.questions import QuestionAnswerValidationError
-from edsl.questions.QuestionBase import QuestionBase
-from edsl.questions.QuestionMatrix import QuestionMatrix
+from edsl.questions import QuestionBase
+from edsl.questions.question_matrix import QuestionMatrix
 from edsl.exceptions.questions import QuestionCreationValidationError
 
 valid_question = {
