+++ tests/questions/test_QuestionNumerical.py
@@ -5,8 +5,8 @@ from edsl.exceptions.questions import (
     QuestionResponseValidationError,
 )
 from edsl.questions import Settings
-from edsl.questions.QuestionBase import QuestionBase
-from edsl.questions.QuestionNumerical import QuestionNumerical  # , main
+from edsl.questions import QuestionBase
+from edsl.questions import QuestionNumerical  # , main
 
 
 valid_question = {
