+++ tests/questions/test_QuestionYesNo.py
@@ -3,8 +3,8 @@ from edsl.exceptions.questions import (
     QuestionResponseValidationError,
     QuestionAnswerValidationError,
 )
-from edsl.questions.QuestionBase import QuestionBase
-from edsl.questions.derived.QuestionYesNo import QuestionYesNo, main
+from edsl.questions import QuestionBase
+from edsl.questions.derived.question_yes_no import QuestionYesNo, main
 
 
 def test_QuestionYesNo_main():
