+++ tests/scenarios/test_Scenario.py
@@ -1,6 +1,6 @@
 import unittest
 from edsl.scenarios import Scenario
-from edsl.questions.QuestionMultipleChoice import QuestionMultipleChoice
+from edsl.questions import QuestionMultipleChoice
 from edsl.jobs import Jobs
 from edsl.agents import Agent
 from edsl.surveys import Survey
